refactor: Move writechain -> sourcenode/writechain
