+++ internal/pkg/service/stream/storage/level/local/writer/sourcenode/writechain/closer.go
